refactor(srv): split subprocess.rs into subprocess/ submodules
Tier 2 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
Bundled controller bookkeeping, a session-id continuity state machine
(deliberately isolated already via a &Mutex<Inner>-not-&self signature), a
host-subprocess turn (spawn_turn, 565 lines, std::process-based), and a
container-exec turn (spawn_container_turn, ~430 lines, bollard/Docker-based)
that never call each other. Split into subprocess/{mod,session,host_spawn,
container_spawn,tests}.rs; both spawn functions moved whole, undecomposed.

Bonus: the 4x-duplicated BlockControllerRuntimeStatus snapshot construction
was factored into one shared build_status_snapshot() helper during the split.

Pure relocation otherwise, no behavior changes. cargo build -p agentmux-srv
clean; cargo test -p agentmux-srv subprocess:: — 13/13 tests pass (all
preserved; corrects an 11-test undercount in the original audit estimate).

<!-- agentmux:agent_id=agentx -->
